+++ src/AddTodoForm.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+function AddTodoForm() {
+    return (
+        <form>
+            <label htmlFor="todoTitle"> Title </label>
+            <br></br>
+            <input id="todoTitle"></input>
+            <button> Add </button>
+        </form>
+    );
+
+}
+
+export default AddTodoForm;
